+++ packages/demo/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.2.6](https://github.com/pie-framework/pie-lib/compare/@pie-lib/demo@8.2.5...@pie-lib/demo@8.2.6) (2025-04-18)
+
+**Note:** Version bump only for package @pie-lib/demo
+
+
+
+
+
+## [8.2.5](https://github.com/pie-framework/pie-lib/compare/@pie-lib/demo@8.2.4...@pie-lib/demo@8.2.5) (2025-04-17)
+
+**Note:** Version bump only for package @pie-lib/demo
+
+
+
+
+
+## [8.2.4](https://github.com/pie-framework/pie-lib/compare/@pie-lib/demo@8.2.3...@pie-lib/demo@8.2.4) (2025-04-17)
+
+**Note:** Version bump only for package @pie-lib/demo
+
+
+
+
+
+## [8.2.3](https://github.com/pie-framework/pie-lib/compare/@pie-lib/demo@8.2.2...@pie-lib/demo@8.2.3) (2025-04-10)
+
+**Note:** Version bump only for package @pie-lib/demo
+
+
+
+
+
 ## [8.2.2](https://github.com/pie-framework/pie-lib/compare/@pie-lib/demo@8.2.1...@pie-lib/demo@8.2.2) (2025-03-31)
 
 **Note:** Version bump only for package @pie-lib/demo